requirements.txt
# The fuel for the rocket 🚀
MetaTrader5
pandas
numpy
python-telegram-bot
gspread
google-api-python-client
google-auth
requests
google-generativeai 
python-dotenv
dotenv
pandas